+++ app/(tabs)/(home)/(touristSpots)/(categories)/[category].tsx
@@ -1,4 +1,6 @@
-// app/(tabs)/(home)/(touristSpots)/(categories)/[category].tsx
+// Category page that displays tourist spots filtered by a specific category
+// Includes search functionality and a list view of spots within the category
+
 import EmptyState from "@/components/EmptyState";
 import { ThemedView } from "@/components/ThemedView";
 import SearchBar from "@/components/touristSpot/TouristSearchBar";
@@ -11,17 +13,17 @@ import {
   StatusBar,
   StyleSheet
 } from "react-native";
-import { TouristSpot, touristSpotsData } from "../data";
-
-type CategoryId = "historical" | "natural" | "urban" | "museums" | "resorts";
+import { TouristSpot, touristSpotsData } from "../TouristSpotData";
 
 const CategoryPage: React.FC = () => {
+  // Get category ID from URL parameters
   const { category: categoryId } = useLocalSearchParams<{ category: string }>();
   const navigation = useNavigation();
   const colorScheme = useColorScheme();
+  // State for search functionality
   const [searchQuery, setSearchQuery] = useState<string>("");
 
-  // Define category title based on categoryId
+  // Helper function to get human-readable category title
   const getCategoryTitle = (id: string): string => {
     switch (id) {
       case "historical":
@@ -39,19 +41,21 @@ const CategoryPage: React.FC = () => {
     }
   };
 
+  // Set navigation header title based on category
   useLayoutEffect(() => {
     navigation.setOptions({
       headerTitle: getCategoryTitle(categoryId),
     });
   }, [navigation, categoryId]);
 
-  // Filter tourist spots based on the categoryId
+  // Filter spots by category
   const allCategorySpots = useMemo(() => {
     return Object.values(touristSpotsData).filter(
       (spot: TouristSpot) => spot.category === categoryId
     );
   }, [categoryId]);
 
+  // Filter spots by search query within the category
   const filteredCategorySpots = useMemo(() => {
     if (!searchQuery.trim()) return allCategorySpots;
 
@@ -61,6 +65,7 @@ const CategoryPage: React.FC = () => {
     );
   }, [searchQuery, allCategorySpots]);
 
+  // Navigation handler for spot selection
   const handleSpotClick = (spotId: string): void => {
     router.push(`/(tabs)/(home)/(touristSpots)/(spots)/${spotId}`);
   };
@@ -71,13 +76,14 @@ const CategoryPage: React.FC = () => {
         barStyle={colorScheme === "dark" ? "light-content" : "dark-content"}
       />
       
+      {/* Search bar for filtering spots within category */}
       <SearchBar
         value={searchQuery}
         onChangeText={setSearchQuery}
         placeholder={`Search ${getCategoryTitle(categoryId).toLowerCase()}...`}
       />
 
-      {/* Content */}
+      {/* List of spots in the category */}
       <ScrollView contentContainerStyle={styles.content}>
         {filteredCategorySpots.length > 0 ? (
           filteredCategorySpots.map((spot) => (
