--- components/TouristSearchBar.tsx
@@ -1,55 +1,0 @@
-import { Colors } from "@/constants/Colors";
-import { useColorScheme } from "@/hooks/useColorScheme";
-import { StyleSheet, TextInput, View } from "react-native";
-
-interface TSearchBarProps {
-  value: string;
-  onChangeText: (text: string) => void;
-  placeholder?: string;
-}
-
-const TSearchBar: React.FC<TSearchBarProps> = ({ value, onChangeText, placeholder }) => {
-  const colorScheme = useColorScheme();
-  const textColor = Colors[colorScheme].text;
-
-  return (
-    <View style={styles.searchContainer}>
-      <View style={styles.searchInputContainer}>
-        <TextInput
-          style={[styles.searchInput, { color: textColor }]}
-          placeholder={placeholder}
-          placeholderTextColor={colorScheme === "dark" ? "#8E9196" : "#9F9EA1"}
-          value={value}
-          onChangeText={onChangeText}
-        />
-      </View>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  searchContainer: {
-    paddingHorizontal: 20,
-    paddingVertical: 10,
-  },
-  searchInputContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-  },
-  searchInput: {
-    flex: 1,
-    backgroundColor: "#fff",
-    borderRadius: 10,
-    padding: 10,
-    paddingHorizontal: 15,
-    fontFamily: "Poppins-Regular",
-    fontSize: 14,
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.25,
-    shadowRadius: 10,
-    elevation: 5,
-  },
-});
-
-export default TSearchBar;
