+++ components/touristSpot/TouristSpotCard.tsx
@@ -1,9 +1,13 @@
+// Reusable card component for displaying tourist spots in different layouts
+// Supports three variants: grid, carousel, and list views
+
 import { ThemedText } from "@/components/ThemedText";
 import { Colors } from "@/constants/Colors";
 import { useColorScheme } from "@/hooks/useColorScheme";
 import { Ionicons } from "@expo/vector-icons";
 import { Image, StyleSheet, TouchableOpacity, View } from "react-native";
 
+// Interface for tourist spot data
 interface TouristSpot {
   image: string;
   name: string;
@@ -22,6 +26,7 @@ const TouristSpotCard: React.FC<TouristSpotCardProps> = ({ spot, onPress, varian
   const cardBackgroundColor = colorScheme === "dark" ? "#1E293B" : "#fff";
   const shadowColor = colorScheme === "dark" ? "#000" : "#ccc";
 
+  // Carousel variant - Large image with overlay text
   if (variant === "carousel") {
     return (
       <TouchableOpacity
@@ -41,6 +46,7 @@ const TouristSpotCard: React.FC<TouristSpotCardProps> = ({ spot, onPress, varian
     );
   }
 
+  // Grid variant - Square image with overlay text
   if (variant === "grid") {
     return (
       <TouchableOpacity
@@ -64,6 +70,7 @@ const TouristSpotCard: React.FC<TouristSpotCardProps> = ({ spot, onPress, varian
     );
   }
 
+  // List variant - Horizontal layout with image and details
   return (
     <TouchableOpacity
       onPress={onPress}
